feat: update database configuration and refactor entity relationships to use Integer IDs

+++ src/main/java/br/gov/agu/nutec/msescala/dto/message/PautaMessage.java
@@ -8,8 +8,8 @@ public record PautaMessage(
         Long pautaId,
         Integer setorOrigemId,
         Integer especieTarefaId,
-        List<Long> avaliadoresIds,
-        List<Long> pautistasIds,
+        List<Integer> avaliadoresIds,
+        List<Integer> pautistasIds,
         String token
 ) {
 }
